Support running standalone with custom options
